citygroves/manager/flesh/pyproject.toml
[tool.poetry]
name = "citygroves.manager"
version = "0.1.0"
description = ""
authors = ["Damian Krystkiewicz <damian.krystkiewicz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0"
Pillow = "^6.2"
loguru = "^0.4.0"
graypy = "^2.1"
django-click = "^2.1"
psycopg2-binary = "^2.8"
djangorestframework = "^3.11"
celery = "^4.3"
django-environ = "^0.4.5"
ipython = "^7.10"
pytest = "^5.3"
pytest-django = "^3.7"
pytest-socket = "^0.3.3"
pytest-env = "^0.6.2"
pyyaml = "^5.2"
redis = "^3.3"
pdfrw = "^0.4.0"
factory_boy = "^2.12"
pytest-factoryboy = "^2.0"
django-extensions = "^2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
